[LLDB][NVIDIA] Remove the initialization stop
I was able to figure out a clean-ish way to remove the initialization
stop and still have our tests pass.
If we use dyld as the stop reason for initialization, LLDB will
autoresume initialization! The dyld stop reason is actually a good one
because it makes sense when doing late attach.
The only change needed in addition to the stop reason is to not wait
synchronously for the gpu target to stop because it won't stop publicly.
I'm going to first send a patch to Greg's branch before commiting this.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

- **Bug Fixes**
	- Improved GPU process connection handling for more reliable debugging sessions.
	- Adjusted GPU thread stop reason messaging during initialization for clearer status indication.
	- Updated breakpoint handling so the GPU process is no longer automatically resumed after a breakpoint is hit.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->