[LLDB][NVIDIA] Move ThreadNVIDIAGPU under ThreadState
Something that hindered some clean information passing was the fact that
the ThreadState objects were detached from the ThreadNVIDIAGPU objects.

Before this patch, ThreadNVIDIAGPU objects are dangling pointers that
get data from ThreadState objects that are dynamically assigned to them.
Not only that, their run/stop state was managed in a manner different
from the batch update logic.

This patches changes that by moving ThreadNVIDIAGPU under ThreadState
and keeping their strong association for the entire life cycle of the
GPU server.

With this, we'll be able to update ThreadNVIDIAGPU as part of batch
updates, which will be very clean.

Not only that, we are now preallocating all ThreadNVIDIAGPU objects
during initialization. They will never be reallocated.

However, I had to do 1 hack. LLGS, the main server interface, wants
the list of threads to be a vector of unique_ptrs. I think changing
that in LLVM upstream is going to be difficult because there's no
compelling reason to do that unless we upstream our own GPU code.
Instead, what I'm doing is simply to create these unique_ptrs but
release the underlying pointer before destruction, so that we never
allocate ThreadNVIDIAGPU after initialization.