[NVIDIA][LLDB] Move warp-level register calculations to WarpState
With this, the WarpState manages the reading of uniform registers.
Invidividual threads can access these registers and have their own copy
to keep LLGS happy.
In the future we need to teach LLGS how to get registers that don't
exist at the thread level, but currently it's very hard to do that.