[LLDB][NVGPU] Skip single steps on breakpoints upon resuming
UD handles that for us, so we need a way to LLDB we need that.
Dmitry said that there's a setting for that, but I really didn't find.
And because we are using the gdb-remote protocol, it seems we need a
custom way to skip this logic.

I'll check with Greg and Dave if they want this patch in the shared
branch or if they prefer a different way to achieve this.