[LLDB][NVGPU] Support two possible libcuda names
On regular linux, libcuda is found as `libcuda.so.1`, but on android it's found as plain `libcuda.so`. When setting up the initialization breakpoint, we need the exact shared library name. The easiest way to handle this is by creating two breakpoints, one on each variant of the name. LLDB will only report one of them during initialization.