Create a trie string table implementation.
After seeing the size of the .debug_str section growing larger and larger over the years, I wanted to see if a trie structured string table would be feasible to implement and use. This patch creates a new TrieStrtab.h and TrieStrtab.cpp which allows encoding and decoding of strings into this new format.

The trie string table is implemented as a trie where the terminal node offsets can be specified to uniquely identify a string. A 32 bit offset can be specified into a section containing the trie and the code can decode a string by reverse traversing the edges up to the root and reconstructing a string.

TrieBuilder is a class that will build a string table in Trie format that is designed to replace the standard string tables from DWARF and object
files. It allows 32 bit offsets to be specified within the Trie that point to the end of the string, and the Trie can be traversed up to the root to
decode the string value.

This allows this string table to be used just like a normal string table where a 32 bit offset is specified, albeit with extra decoding needed to get
the re-create the string values. During testing of some large C++ binaries, the trie string table was 45% of the size of the original string table. A
DWARF optimizer could modify existing DWARF to switch over to using this new string table without having to re-write the stream of bytes in DWARF since
a 32 bit string table offsets for both tables are the same size. A new DWARF form would need to be added to support this. Something like:

 DW_FORM_strp_trie

So a DWARF optimizer could easily re-write the .debug_abbrev section to replace any DW_FORM_strp entries with DW_FORM_strp_trie, replace the existing .debug_str section a new optimized .debug_str_trie and then fixup all previous DW_FORM_strp values with updated DW_FORM_strp_trie offsets.