[LLDB][NVGPU] Sync the two processes conditionally
Lately we removed synchronization between the two proceses to simplify
test orchestration because we are not handling shadow breakpoints yet,
which can cause spurious stop on the cpu after setting breakpoints
intended for the gpu. However, this is not great for CLI usage. LLDB
currently doesn't let you run commands unless both processes are
stopped. You could cltr-c on every stop to force both processes to be
stopped, but that's not ideal.
Instead, I'm adding an env var that controls this behavior so that the
CLI works as before, and tests can use this as a workaround until we
have shadow breakpoints.
The behavior of the CLI is something we'll need to fix if we want true
non-stop mode between these two processes in the CLI.
On VSCode, this won't be a problem.