Add tracking old offset to new offset.
Modified the TrieBuilder to track the original offset to new offset mapping. This allows anyone the create a string table using the TrieBuilder and then have a map between the old offset from the .debug_str section or ELF string table, to the new offset in the trie string table.