[GPU][AMD] Add support for address space reads (#67)
This commit enables the amd gpu process to read from different address spaces. We the following address spaces:

  * generic - flat address space
  * region - gds address space (not supported on MI300/MI350)
  * local - threadgroup shared memory address space
  * private_lane - thread local address space
  * private_wave - wave local address space

These address spaces correspond to address spaces exposed for dwarf expressions.

In order to read from the private_lane address space we need to keep the lane as part of the ThreadAMDGPU object. This is a straightforward modification to the existing code that creates the per-lane threads.

Tests were added to exercise each of the exposed address spaces.